+++ src/api/index.ts
@@ -1,3 +1,8 @@
-export * from './loadSupport'
-export * from './runCucumber'
+/*
+Anything exported here will be publicly available via `@cucumber/cucumber/api`
+ */
+
+export * from './load_configuration'
+export * from './load_support'
+export * from './run_cucumber'
 export * from './types'