+++ src/configuration/argv_parser.ts
@@ -1,64 +1,36 @@
 import { Command } from 'commander'
+import merge from 'lodash.merge'
 import path from 'path'
 import { dialects } from '@cucumber/gherkin'
-import { SnippetInterface } from '../formatter/step_definition_snippet_builder/snippet_syntax'
 import Formatters from '../formatter/helpers/formatters'
 import { version } from '../version'
-import { PickleOrder } from './helpers'
-
-export interface IParsedArgvFormatRerunOptions {
-  separator?: string
-}
-
-export interface IParsedArgvFormatOptions {
-  colorsEnabled?: boolean
-  rerun?: IParsedArgvFormatRerunOptions
-  snippetInterface?: SnippetInterface
-  snippetSyntax?: string
-  printAttachments?: boolean
-  [customKey: string]: any
-}
+import { IConfiguration } from './types'
 
 export interface IParsedArgvOptions {
-  backtrace: boolean
-  config: string
-  dryRun: boolean
-  exit: boolean
-  failFast: boolean
-  format: string[]
-  formatOptions: IParsedArgvFormatOptions
-  i18nKeywords: string
-  i18nLanguages: boolean
-  import: string[]
-  language: string
-  name: string[]
-  order: PickleOrder
-  parallel: number
+  config?: string
+  i18nKeywords?: string
+  i18nLanguages?: boolean
   profile: string[]
-  publish: boolean
-  publishQuiet: boolean
-  require: string[]
-  requireModule: string[]
-  retry: number
-  retryTagFilter: string
-  strict: boolean
-  tags: string
-  worldParameters: object
 }
 
 export interface IParsedArgv {
-  args: string[]
   options: IParsedArgvOptions
+  configuration: Partial<IConfiguration>
 }
 
+type IRawArgvOptions = Partial<Omit<IConfiguration, 'paths'>> &
+  IParsedArgvOptions
+
 const ArgvParser = {
-  collect<T>(val: T, memo: T[]): T[] {
-    memo.push(val)
-    return memo
+  collect<T>(val: T, memo: T[] = []): T[] {
+    if (val) {
+      return [...memo, val]
+    }
+    return undefined
   },
 
-  mergeJson(option: string): (str: string, memo: object) => object {
-    return function (str: string, memo: object) {
+  mergeJson(option: string): (str: string, memo?: object) => object {
+    return function (str: string, memo: object = {}) {
       let val: object
       try {
         val = JSON.parse(str)
@@ -69,12 +41,12 @@ const ArgvParser = {
       if (typeof val !== 'object' || Array.isArray(val)) {
         throw new Error(`${option} must be passed JSON of an object: ${str}`)
       }
-      return { ...memo, ...val }
+      return merge(memo, val)
     }
   },
 
-  mergeTags(value: string, memo: string): string {
-    return memo === '' ? `(${value})` : `${memo} and (${value})`
+  mergeTags(value: string, memo?: string): string {
+    return memo ? `${memo} and (${value})` : `(${value})`
   },
 
   validateCountOption(value: string, optionName: string): number {
@@ -92,14 +64,6 @@ const ArgvParser = {
     return value
   },
 
-  validateRetryOptions(options: IParsedArgvOptions): void {
-    if (options.retryTagFilter !== '' && options.retry === 0) {
-      throw new Error(
-        'a positive --retry count must be specified when setting --retry-tag-filter'
-      )
-    }
-  },
-
   parse(argv: string[]): IParsedArgv {
     const program = new Command(path.basename(argv[1]))
 
@@ -108,60 +72,48 @@ const ArgvParser = {
       .usage('[options] [<GLOB|DIR|FILE[:LINE]>...]')
       .version(version, '-v, --version')
       .option('-b, --backtrace', 'show full backtrace for errors')
-      .option('-c, --config <TYPE[:PATH]>', 'specify configuration file')
+      .option('-c, --config <PATH>', 'specify configuration file')
+      .option('-d, --dry-run', 'invoke formatters without executing steps')
       .option(
-        '-d, --dry-run',
-        'invoke formatters without executing steps',
-        false
+        '--exit, --force-exit',
+        'force shutdown of the event loop when the test run has finished: cucumber will call process.exit'
       )
-      .option(
-        '--exit',
-        'force shutdown of the event loop when the test run has finished: cucumber will call process.exit',
-        false
-      )
-      .option('--fail-fast', 'abort the run on first failure', false)
+      .option('--fail-fast', 'abort the run on first failure')
       .option(
         '-f, --format <TYPE[:PATH]>',
         'specify the output format, optionally supply PATH to redirect formatter output (repeatable).  Available formats:\n' +
           Formatters.buildFormattersDocumentationString(),
-        ArgvParser.collect,
-        []
+        ArgvParser.collect
       )
       .option(
         '--format-options <JSON>',
         'provide options for formatters (repeatable)',
-        ArgvParser.mergeJson('--format-options'),
-        {}
+        ArgvParser.mergeJson('--format-options')
       )
       .option(
         '--i18n-keywords <ISO 639-1>',
         'list language keywords',
-        ArgvParser.validateLanguage,
-        ''
+        ArgvParser.validateLanguage
       )
-      .option('--i18n-languages', 'list languages', false)
+      .option('--i18n-languages', 'list languages')
       .option(
-        '--import <GLOB|DIR|FILE>',
+        '-i, --import <GLOB|DIR|FILE>',
         'import files before executing features (repeatable)',
-        ArgvParser.collect,
-        []
+        ArgvParser.collect
       )
       .option(
         '--language <ISO 639-1>',
-        'provide the default language for feature files',
-        'en'
+        'provide the default language for feature files'
       )
       .option(
         '--name <REGEXP>',
         'only execute the scenarios with name matching the expression (repeatable)',
-        ArgvParser.collect,
-        []
+        ArgvParser.collect
       )
-      .option('--no-strict', 'succeed even if there are pending steps')
+
       .option(
         '--order <TYPE[:SEED]>',
-        'run scenarios in the specified order. Type should be `defined` or `random`',
-        'defined'
+        'run scenarios in the specified order. Type should be `defined` or `random`'
       )
       .option(
         '-p, --profile <NAME>',
@@ -172,55 +124,45 @@ const ArgvParser = {
       .option(
         '--parallel <NUMBER_OF_WORKERS>',
         'run in parallel with the given number of workers',
-        (val) => ArgvParser.validateCountOption(val, '--parallel'),
-        0
+        (val) => ArgvParser.validateCountOption(val, '--parallel')
       )
-      .option(
-        '--publish',
-        'Publish a report to https://reports.cucumber.io',
-        false
-      )
+      .option('--publish', 'Publish a report to https://reports.cucumber.io')
       .option(
         '--publish-quiet',
-        "Don't print information banner about publishing reports",
-        false
+        "Don't print information banner about publishing reports"
       )
       .option(
         '-r, --require <GLOB|DIR|FILE>',
         'require files before executing features (repeatable)',
-        ArgvParser.collect,
-        []
+        ArgvParser.collect
       )
       .option(
         '--require-module <NODE_MODULE>',
         'require node modules before requiring files (repeatable)',
-        ArgvParser.collect,
-        []
+        ArgvParser.collect
       )
       .option(
         '--retry <NUMBER_OF_RETRIES>',
         'specify the number of times to retry failing test cases (default: 0)',
-        (val) => ArgvParser.validateCountOption(val, '--retry'),
-        0
+        (val) => ArgvParser.validateCountOption(val, '--retry')
       )
       .option(
         '--retry-tag-filter <EXPRESSION>',
         `only retries the features or scenarios with tags matching the expression (repeatable).
         This option requires '--retry' to be specified.`,
-        ArgvParser.mergeTags,
-        ''
+        ArgvParser.mergeTags
       )
+      .option('--strict', 'fail if there are pending steps')
+      .option('--no-strict', 'succeed even if there are pending steps')
       .option(
         '-t, --tags <EXPRESSION>',
         'only execute the features or scenarios with tags matching the expression (repeatable)',
-        ArgvParser.mergeTags,
-        ''
+        ArgvParser.mergeTags
       )
       .option(
         '--world-parameters <JSON>',
         'provide parameters that will be passed to the world constructor (repeatable)',
-        ArgvParser.mergeJson('--world-parameters'),
-        {}
+        ArgvParser.mergeJson('--world-parameters')
       )
 
     program.addHelpText(
@@ -229,12 +171,26 @@ const ArgvParser = {
     )
 
     program.parse(argv)
-    const options: IParsedArgvOptions = program.opts()
-    ArgvParser.validateRetryOptions(options)
+    const {
+      config,
+      i18nKeywords,
+      i18nLanguages,
+      profile,
+      ...regularStuff
+    }: IRawArgvOptions = program.opts()
+    const configuration: Partial<IConfiguration> = regularStuff
+    if (program.args.length > 0) {
+      configuration.paths = program.args
+    }
 
     return {
-      options,
-      args: program.args,
+      options: {
+        config,
+        i18nKeywords,
+        i18nLanguages,
+        profile,
+      },
+      configuration,
     }
   },
 }