+++ src/configuration/types.ts
@@ -1,41 +1,26 @@
-import { IRuntimeOptions } from '../runtime'
-import { IParsedArgvFormatOptions } from '../cli/argv_parser'
-import { PickleOrder } from '../cli/helpers'
-import {
-  ISupportCodeCoordinates,
-  ISupportCodeLibrary,
-} from '../support_code_library_builder/types'
+import { FormatOptions } from '../formatter'
+import { PickleOrder } from '../models/pickle_order'
 
-export interface ISourcesCoordinates {
-  defaultDialect?: string
-  paths?: string[]
-  names?: string[]
-  tagExpression?: string
-  order?: PickleOrder
-}
-
-export interface IUserConfiguration {
-  sources: ISourcesCoordinates
-  support: ISupportCodeCoordinates
-  runtime?: Partial<IRuntimeOptions> & { parallel?: number }
-  formats?: IFormatterConfiguration
-}
-
-export interface IRunConfiguration {
-  sources: ISourcesCoordinates
-  support: ISupportCodeCoordinates | ISupportCodeLibrary
-  runtime?: Partial<IRuntimeOptions> & { parallel?: number }
-  formats?: IFormatterConfiguration
-}
-
-export interface IFormatterConfiguration {
-  stdout?: string
-  files?: Record<string, string>
-  publish?:
-    | {
-        url?: string
-        token?: string
-      }
-    | false
-  options?: IParsedArgvFormatOptions
+export interface IConfiguration {
+  backtrace: boolean
+  dryRun: boolean
+  forceExit: boolean
+  failFast: boolean
+  format: string[]
+  formatOptions: FormatOptions
+  import: string[]
+  language: string
+  name: string[]
+  order: PickleOrder
+  paths: string[]
+  parallel: number
+  publish: boolean
+  publishQuiet: boolean
+  require: string[]
+  requireModule: string[]
+  retry: number
+  retryTagFilter: string
+  strict: boolean
+  tags: string
+  worldParameters: any
 }